EE-1363: added "clean" sub-command and "--delete-temp-files" command-line argument.
Increased the test coverage to 97% (from 85%).

Removed all the 338 errors reported by pycodestyle.

Reduced the 630 errors (reported by pylint) to 36
(when the reporting for smaller errors
(like missing-function-docstring) are turned off).

Reduced the 702 errors (reported by pylint) to 149
(when the reporting for smaller errors
(like missing-function-docstring) are turned on).

Removed all the global variables.

Rewrote the program in object-oriented style.

Created a test which makes it possible to test the scenario
when the client receives a "Connection reset by the server" error.

Added an .editorconfig file to prevent mixing tabs with spaces.

Added a .gitignore file.

Breaking the build if "coverage", "pylint" or "pycodestyle" aren't satisfied.